Relax SFW filtering for VNDB
Previously entries with any NSFW releases would be filtered.
This has been relaxed to allow entries with either one standalone
SFW release, zero standalone NSFW releases, or the
`In-game Sexual Content Toggle` tag.

The `GameModel.image` field is now rudimentarily expunged of sensitive
images when the SFW filter is enabled. This is not ideal, but likely
requires `GameModel` changes to address properly.